Se realiza el ajuste a nivel de modelo y dao Usuario con el ajuste de la base de datos, también se elimina la falla que se presentaba con anterioridad en sprint-framework

+++ Fuentes/ServiceVentaComida/src/main/java/co/edu/usbcali/dao/IUsuarioDAO.java
@@ -9,7 +9,7 @@ public interface IUsuarioDAO {
 	public void crear(Usuario entity);
 	public void modificar(Usuario entity);
 	public void borrar(Usuario entity);
-	public Usuario consultarPorId(int usuId);
+	public Usuario consultarPorId(long usuId);
 	public Usuario consultarPorIdentificacion(Integer numIdentificacion, Integer tipoIdentificacion);
 	public List<Usuario> consultarTodos();
